+++ src/ui/buttons/buttons.jsx
@@ -1,5 +1,6 @@
 /* @flow */
 /** @jsx node */
+
 import { node, type ElementNode } from 'jsx-pragmatic/src';
 import { FUNDING, WALLET_INSTRUMENT } from '@paypal/sdk-constants/src';
 import { noop } from 'belter/src';
@@ -9,7 +10,8 @@ import { CLASS, BUTTON_NUMBER, BUTTON_LAYOUT, BUTTON_FLOW } from '../../constant
 import { determineEligibleFunding, isWalletFundingEligible } from '../../funding';
 import { ValidationError } from '../../lib';
 
-import { getButtonAnimation } from './button-animations';
+import { getButtonDesign } from './buttonDesigns';
+import { ButtonDesignExperimentScriptWrapper } from './buttonDesigns/script';
 import { normalizeButtonProps, type ButtonPropsInputs, type OnShippingChange } from './props';
 import { Style } from './style';
 import { Button } from './button';
@@ -108,7 +110,7 @@ export function Buttons(props : ButtonsProps) : ElementNode {
     }
 
     if (fundingSources.indexOf(FUNDING.CARD) !== -1) {
-        fundingSources = fundingSources.filter(src => src !== FUNDING.CARD).concat([ FUNDING.CARD ]);
+        fundingSources = [ ...fundingSources.filter(src => src !== FUNDING.CARD),  FUNDING.CARD ];
     }
 
     const instruments = getWalletInstruments({ wallet, fundingSources, layout, onShippingChange });
@@ -117,7 +119,9 @@ export function Buttons(props : ButtonsProps) : ElementNode {
         flow === BUTTON_FLOW.PURCHASE &&
         ((__WEB__ && userIDToken) || Object.keys(instruments).length)
     );
+
     const buttonAnimation = getButtonAnimation(personalizations);
+
     return (
         <div class={ [
             CLASS.CONTAINER,
@@ -138,7 +142,6 @@ export function Buttons(props : ButtonsProps) : ElementNode {
             {
                 fundingSources.map((source, i) => (
                     <Button
-                        buttonAnimation={ source && source === FUNDING.PAYPAL ? buttonAnimation : null }
                         content={ content }
                         i={ i }
                         style={ style }
@@ -184,8 +187,8 @@ export function Buttons(props : ButtonsProps) : ElementNode {
             {
                 (layout === BUTTON_LAYOUT.VERTICAL && fundingSources.indexOf(FUNDING.CARD) !== -1)
                     ? <PoweredByPayPal
-                        locale={ locale }
-                        nonce={ nonce }
+                            locale={ locale }
+                            nonce={ nonce }
                     /> : null
             }