+++ src/ui/buttons/props.js
@@ -272,7 +272,8 @@ export type ButtonProps = {|
     supportedNativeBrowser : boolean,
     applePaySupport : boolean,
     applePay : ApplePaySessionConfigRequest,
-    meta : {||}
+    meta : {||},
+    renderedButtons : $ReadOnlyArray<$Values<typeof FUNDING>>
 |};
 
 // eslint-disable-next-line flowtype/require-exact-type